backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "Backend for Campus Connect"
authors = ["hemanth18 <hg510443@gmail.com>"]

package-mode = false

[tool.poetry.dependencies]
python = ">=3.12"
fastapi = ">=0.115.12,<0.116.0"
uvicorn = { extras = ["standard"], version = ">=0.34.3,<0.35.0" }
sqlalchemy = ">=2.0.41,<3.0.0"
alembic = ">=1.16.1,<2.0.0"
python-dotenv = ">=1.1.0,<2.0.0"
pydantic = {extras = ["email"], version = "^2.11.5"}
passlib = ">=1.7.4,<2.0.0"
bcrypt = ">=3.2.0,<4.0.0"
python-jose = "^3.5.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
